DFR-1707: Refactor test classes
