+++ src/test/java/uk/gov/hmcts/reform/finrem/caseorchestration/model/ccd/spreadsheet/CCDConfigValidator.java
@@ -8,7 +8,6 @@ import org.apache.poi.ss.usermodel.Row;
 import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.xssf.usermodel.XSSFWorkbook;
-import uk.gov.hmcts.reform.finrem.caseorchestration.model.ccd.AccessingAllClassesInPackage;
 
 import java.io.File;
 import java.io.IOException;
